core-model/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "prism-core"
version = "0.4.0"
description = "PRISM sizing engine"
requires-python = ">=3.12"
dependencies = [
    "prism-common>=0.1.0",
    "pydantic>=2.7",
    "PyYAML>=6",
    "typer[all]>=0.12",
    "asteval>=0.9",
    "jsonschema>=4.22",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["bims.prism*"]
exclude = ["bims.prism.adapters*"]  # ядру не нужны адаптеры
